% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ax}
\alias{ax}
\title{12000 strings}
\format{
A data frame with 1000 rows and 12 variables
\describe{
  \item{Code}{string}
}
}
\usage{
a
}
\description{
The dataset was created from data available at \url{https://www.prescient.pro/}.
}
\keyword{datasets}
