% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{desc}
\alias{desc}
\title{Function to calculate descriptive statistics for a variable}
\usage{
desc(variable, name)
}
\arguments{
\item{variable}{is a vector of the variable}

\item{name}{is the name of the variable}
}
\value{
a vector with a list of descriptive statistics
}
\description{
Function to calculate descriptive statistics for a variable
}
\keyword{descriptive,}
\keyword{statistics}
