% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Function to calculate descriptive statistics for a vector variable}
\usage{
descriptive(variable, name)
}
\arguments{
\item{variable}{is a vector of the variable}

\item{name}{is the name of the variable}
}
\value{
a vector with a list of descriptive statistics
}
\description{
Function to calculate descriptive statistics for a vector variable
}
\keyword{descriptive,}
\keyword{statistics}
