% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_categorical_feature_dist.R
\name{df_categorical_feature_dist}
\alias{df_categorical_feature_dist}
\title{Calculate the feature-level PSI index for the specified numeric features}
\usage{
df_categorical_feature_dist(df, gather_cols, key = "key", value = "value")
}
\arguments{
\item{df}{tbd}

\item{gather_cols}{tbd}

\item{key}{tbd}

\item{value}{tbd}
}
\value{
base R dataframe
}
\description{
This function takes two Spark DataFrames as input.  One should contain the features
with their expected values.  The other should contain the features with their actual
values.  Example... if we're comparing Oct '18 to Nov '18 features, Oct '18 would be
expected and Nov '18 would be actual.
}
\details{
NOTE: This function currently only supports NUMERIC and/or CHARACTER data types.
}
