% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_model_monitoring_magnitude.R
\name{df_cramers_v_magnitude}
\alias{df_cramers_v_magnitude}
\title{Identify the effect and magnitude of CramersVScore for Chi Square goodness of fit tests}
\usage{
df_cramers_v_magnitude(CramersVScore, DegreesOfFreedom)
}
\arguments{
\item{CramersVScore}{Required: The Cramer's V score for goodness of fit.}

\item{DegreesOfFreedom}{Required: degrees of freedom from the ChiSquared test / number of bins - 1}
}
\value{
The magnitude of Cramer's V score: negligible, low, moderate, or large
}
\description{
Given a cramer's V score and degrees of freedom, this function will return one of
four magnitudes to help identify whether a give feature's distribution has drifted
from it's baseline.
}
\details{
This function is called based on the assumption that a Chi Square goodness of fit has
already been performed, as well as a Cramer's V test, and that the v score and
degrees of freedom are available.
}
