% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_get_feature_distribution.R
\name{df_get_feature_distribution}
\alias{df_get_feature_distribution}
\title{Calculate the bin-level PSI index for the specified numeric features}
\usage{
df_get_feature_distribution(expected_, actual_, features_)
}
\arguments{
\item{expected_}{Required: A matrix containing features with the expected (old) data.}

\item{actual_}{Required: A matrix containing features from with the actual (new) data.}

\item{features_}{Optional: A vector of the feature names to validate.  Note, the feature names must exist in both expected_ and actual_ and be of the same data type in each data frame.  If not features are provided, all features in expected_ will be used.}
}
\value{
A matrix containing the feature name, bin, min value, max value, expected count, expected %, actual count, actual %, and index
}
\description{
This function takes two matrices as input.  One should contain the features
with their expected values.  The other should contain the features with their actual
values.  Example... if we're comparing Oct '18 to Nov '18 features, Oct '18 would be
expected and Nov '18 would be actual.
}
\details{
NOTE: This function currently only supports NUMERIC and/or CHARACTER datatypes
}
