% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_get_psi_score.R
\name{df_get_psi_score}
\alias{df_get_psi_score}
\title{Calculate the feature-level PSI index for the specified numeric features}
\usage{
df_get_psi_score(expected_, actual_, features_)
}
\arguments{
\item{expected_}{Required: A matrix containing features with the expected (old) data.}

\item{actual_}{Required: A matrix containing features from with the actual (new) data.}

\item{features_}{Optional: A vector of the feature names to validate.  Note, the feature names must exist in both sdf_expected and sdf_actual and be of the same data type in each data frame.  If not features are provided, all features in sdf_expected will be used.}
}
\value{
A matrix containing the feature name, bin, min value, max value, expected count, expected %, actual count, actual %, and index
}
\description{
This function takes two Spark DataFrames as input.  One should contain the features
with their expected values.  The other should contain the features with their actual
values.  Example... if we're comparing Oct '18 to Nov '18 features, Oct '18 would be
expected and Nov '18 would be actual.
}
\details{
NOTE: This function currently only supports NUMERIC and/or CHARACTER data types. If you have
      other types of data, please filter them out before passing to the function.
}
