% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_means.R
\name{df_means}
\alias{df_means}
\title{Get the means of the numeric columns in a dataframe}
\usage{
df_means(df)
}
\arguments{
\item{df}{a base R dataframe contain features to be averaged.If non-numeric features are included they will be removed before averaging.}
}
\value{
A base R dataframe containing one row per feature with the feature name and mean
}
\description{
Get the means of the numeric columns in a dataframe
}
