% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_model_monitoring_summary.R
\name{df_model_monitoring_summary}
\alias{df_model_monitoring_summary}
\title{Assess observed and expected feature distributions for drift.}
\usage{
df_model_monitoring_summary(distributions)
}
\arguments{
\item{distributions}{Required: Distributions for the features.  Can be created via the df_model_monitoring_distributions function.}
}
\value{
A base R dataframe containing the feature, # of bins, Chi Squared Value, P-Value, Cramer's V, Result, and Magnitude.
}
\description{
This function will take a distribution table of observed and expected features
and perform a Chi Square Goodness of Fit test for all numeric and categorical
features.  Furthermore, with will account for sample size and degrees of freedom
by performing a subsequent Cramers V Goodness of Fit test.
}
