% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_validate.R
\name{df_validate}
\alias{df_validate}
\title{Validate two separate datasets and return validation data on aggregate and id level features}
\usage{
df_validate(actual, expected, features, key, matchTestPct = 0.98,
  meanTestPct = 0.02)
}
\arguments{
\item{actual}{a base R dataframe contain actual feature values}

\item{expected}{a base R dataframe containing expected feature values}

\item{features}{the features to be validated}

\item{key}{The name of the primary key for actual and expected}

\item{matchTestPct}{The percentage of exact, id level matches to pass test 1}

\item{meanTestPct}{The maximimun percent difference allowed between actual and expected features}
}
\value{
A base R dataframe containing one row per feature with match and mean validation data
}
\description{
This function takes two datasets (actual and expected) and validates them based on how many
features match exact values at the id level as well as an aggregate % difference in the means.
}
\details{
NOTE: The columns to be validate must have the same name in each dataset
}
