% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.tutorial-mitPasswort.R, R/ds.tutorial.R
\name{ds.tutorial}
\alias{ds.tutorial}
\title{Function to start a tutorial accompanying the package "datascience" and the book "Data Science und Statistik mit R"}
\usage{
ds.tutorial(name)

ds.tutorial(name)
}
\arguments{
\item{name}{is the name of the tutorial}
}
\description{
Function to start a tutorial accompanying the package "datascience" and the book "Data Science und Statistik mit R"
}
\keyword{tutorial}
