% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots number of accidents as points for selected State and Year}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A charecter string or Integer of state number of USA}

\item{year}{A character string or Integer of Year}
}
\value{
Plots graph displaying accidents as points for selected state and year
 Function will throw error incase of state number not found in dataset for
 selected year.
}
\description{
This function draws accident data for state and accidents are
represented as points and state boundry is displayed by line.
Location of accident is identified by range of Longitude and latitude but NA
values are ignored to pricise points of accidents.
}
\examples{
\dontrun{
fars_map_state(5,2015)
fars_map_state(5,"2015")
fars_map_state("5",2015)
fars_map_state("5","2015")
}

}

