% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Imports CSV file and convert it to A Dataframe table}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A charecter string to provide filename}
}
\value{
This function returns Dataframe table of imported CSV data
}
\description{
This is simple function to import csv data and return dataframe 
table for data operations. Function will fail if filename does't
exist in working directory.
}
\examples{
fars_read("accident_2013.csv.bz2")

}

