% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparklyr_PSI_Functions.r
\name{get_psi_score}
\alias{get_psi_score}
\title{Calculate the feature-level PSI index for the specified numeric features}
\usage{
get_psi_score(sdf_expected, sdf_actual, features)
}
\arguments{
\item{sdf_expected}{A Spark DataFrame containing features with the expected (old) data.}

\item{features}{A vector of the numeric feature names to validate.  Note, the feature names must exist in both sdf_expected and sdf_new and the features must be numeric in each DataFrame}

\item{sdf_new}{A Spark DataFrame containing features from with the actual (new) data.}
}
\value{
A tbl_Spark containing the feature name and PSI score
}
\description{
This function takes two Spark DataFrames as input.  One should contain the features
with their expected values.  The other should contain the features with their actual
values.  Example... if we're comparing Oct '18 to Nov '18 features, Oct '18 would be
expected and Nov '18 would be actual.
}
\details{
NOTE: This function currently only supports NUMERIC features.  Categorical and Date features will
be added in the future.
}
