% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Creates filename based on provided Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A integer or character string of Year}
}
\value{
This functio nreturns charecter string of Filename
}
\description{
This function is used to generate filename from provided year in
argument. Accident Data is organized per year and compressed in different
files with specific pattern. This function simplifized user input by just
providing Year and compose filename.
}
\examples{
\dontrun{
make_filename("2015")
make_filename(2014)
}

}

