% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_bins.R
\name{sdf_bins}
\alias{sdf_bins}
\title{Calculate the bin-level PSI index for the specified numeric features}
\usage{
sdf_bins(sdf, gather_cols)
}
\arguments{
\item{sdf_expected}{Required: A Spark DataFrame containing features with the expected (old) data.}

\item{sdf_new}{Required: A Spark DataFrame containing features from with the actual (new) data.}

\item{features}{Optional: A vector of the feature names to validate.  Note, the feature names must exist in both sdf_expected and sdf_actual and be of the same data type in each data frame.  If not features are provided, all features in sdf_expected will be used.}

\item{bins}{Optional: An int (example, 10L) value representing the number of bins to create for the continuous variables.  Actuall bins may be less depending on the distribution}
}
\value{
A tbl_Spark containing the feature name, bin, min value, max value, expected count, expected %, actual count, actual %, and index
}
\description{
This function takes two Spark DataFrames as input.  One should contain the features
with their expected values.  The other should contain the features with their actual
values.  Example... if we're comparing Oct '18 to Nov '18 features, Oct '18 would be
expected and Nov '18 would be actual.
}
\details{
NOTE: This function currently only supports NUMERIC and/or CHARACTER datatypes
}
