% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf_get_feature_bins.R
\name{sdf_get_feature_bins}
\alias{sdf_get_feature_bins}
\title{Create bins for numeric features}
\usage{
sdf_get_feature_bins(sdf, features, bins)
}
\arguments{
\item{sdf}{A Spark DataFrame containing features that need to be binned.}

\item{features}{A vector of the numeric feature names to be binned}

\item{bins}{The number of bins to create in.  Use the L suffix to ensure the value is an integer.}
}
\value{
A tbl_Spark containing the feature name, bin number, min value, and max value
}
\description{
This function takes a Spark DataFrame and returns a Spark DataFrame with the designated number
of bins for the features.  Depending on the range and distribution of the numeric feature,
this function may return less than the designated number of bins.
}
\details{
NOTE:  This function currently only supports NUMERIC features.  Categorical and Date features will
be added in the future.
}
