% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_get_feature_bins.R
\name{df_get_feature_bins}
\alias{df_get_feature_bins}
\title{Create bins for numeric features}
\usage{
df_get_feature_bins(data_, features_ = NULL, dataType = "numeric",
  bins_ = NULL)
}
\arguments{
\item{data_}{A data.frame containing features that need to be binned.}

\item{features_}{A vector of the numeric feature names to be binned}

\item{bins}{The number of bins to create in.  Use the L suffix to ensure the value is an integer.}
}
\value{
A data.frame containing the feature name, bin number, min value, and max value
}
\description{
This function takes a matrix and returns a matrix with the designated number
of bins for the features.  Depending on the range and distribution of the numeric feature,
this function may return less than the designated number of bins.
}
\details{
NOTE: This function currently only supports NUMERIC and/or CHARACTER data types. If you have
      other types of data, please filter them out before passing to the function.
}
