% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.summary.R
\name{ds.summary}
\alias{ds.summary}
\title{Function to calculate descriptive statistics for a variable}
\usage{
ds.summary(variable, titel = "", achse = "")
}
\arguments{
\item{variable}{is a vector of the variable}

\item{titel}{is the name of the variable to be used as title on diagrams}

\item{achse}{is the description of the axis on diagrams}
}
\value{
a vector with a list of descriptive statistics
}
\description{
Function to calculate descriptive statistics for a variable
}
\keyword{descriptive,}
\keyword{statistics,}
\keyword{summary}
