% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sparklyr_PSI_Functions.r
\name{get_feature_distribution}
\alias{get_feature_distribution}
\title{Calculate the bin-level PSI index for the specified numeric features}
\usage{
get_feature_distribution(sdf_expected, sdf_actual, features = NULL,
  bins = NULL)
}
\arguments{
\item{sdf_expected}{Required: A Spark DataFrame containing features with the expected (old) data.}

\item{features}{Optional: A vector of the numeric feature names to validate.  Note, the feature names must exist in both sdf_expected and sdf_new and the features must be numeric in each DataFrame}

\item{bins}{Optional: An int (example, 10L) value representing the number of bins to create for the continuous variables.  Actuall bins may be less depending on the distribution}

\item{sdf_new}{Required: A Spark DataFrame containing features from with the actual (new) data.}
}
\value{
A tbl_Spark containing the feature name, bin, min value, max value, expected count, expected %, actual count, actual %, and index
}
\description{
This function takes two Spark DataFrames as input.  One should contain the features
with their expected values.  The other should contain the features with their actual
values.  Example... if we're comparing Oct '18 to Nov '18 features, Oct '18 would be
expected and Nov '18 would be actual.
}
\details{
NOTE: This function currently only supports NUMERIC features.  Categorical and Date features will
be added in the future.
}
